requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
opensearch-py
sentence-transformers
faiss-cpu
numpy
pandas
plotly
httpx
python-multipart
reportlab
